+++ lib/contributions.ml
@@ -2,9 +2,10 @@ module Json = Yojson.Safe
 
 let ( / ) a b = Json.Util.member b a
 
-let query =
-  {|query($from: DateTime!, $to: DateTime!) {
-   viewer {
+let query user =
+  Format.sprintf
+    {|query($from: DateTime!, $to: DateTime!) {
+   %s {
     login
     contributionsCollection(from: $from, to: $to) {
       issueContributions(first: 100) {
@@ -54,9 +55,13 @@ let query =
     }
   }
 }|}
+    (match user with
+    | Some u -> Format.sprintf "user(login: %S)" u
+    | None -> "viewer")
 
-let request ~period:(start, finish) ~token =
+let request ~period:(start, finish) ~user ~token =
   let variables = [ ("from", `String start); ("to", `String finish) ] in
+  let query = query user in
   Graphql.request ~token ~variables ~query ()
 
 module Datetime = struct
@@ -137,9 +142,19 @@ let read_repos json =
        repo;
      }
 
-let of_json ~from json =
-  let username = json / "data" / "viewer" / "login" |> Json.Util.to_string in
-  let contribs = json / "data" / "viewer" / "contributionsCollection" in
+let of_json ~from ~user json =
+  let username, contribs =
+    match user with
+    | Some username ->
+        let contribs = json / "data" / "user" / "contributionsCollection" in
+        (username, contribs)
+    | None ->
+        let username =
+          json / "data" / "viewer" / "login" |> Json.Util.to_string
+        in
+        let contribs = json / "data" / "viewer" / "contributionsCollection" in
+        (username, contribs)
+  in
   let items =
     read_issues (contribs / "issueContributions")
     @ read_prs (contribs / "pullRequestContributions")